requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amazon-textract-caller==0.2.3
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.7.9
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.34.84
botocore==1.34.84
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
doctran==0.0.14
docx2txt==0.8
editdistance==0.8.1
effdet==0.4.1
email_validator==2.1.1
emoji==2.11.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.13.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.6
intel-openmp==2021.4.0
iopath==0.1.10
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.41
langchain-openai==0.1.2
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.43
layoutparser==0.3.4
lxml==4.9.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.1
openai==0.27.10
opencv-python==4.9.0.80
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
phonenumbers==8.13.34
pikepdf==8.15.1
pillow==10.3.0
pillow_heif==0.16.0
portalocker==2.8.2
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.16.3
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.29.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.9.0
rapidocr-onnxruntime==1.3.16
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scipy==1.13.0
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.5.2
timm==0.9.16
tldextract==5.1.2
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
